orginize the folderes
