+++ frontend/rule-managment/src/App.css
@@ -26,7 +26,8 @@ table {
   border-collapse: collapse;
 }
 
-th, td {
+th,
+td {
   padding: 10px;
   border-bottom: 1px solid #ddd;
   text-align: left;
